reintroduce wiping of db by env var

+++ scripts/launch-indexer.sh
@@ -22,6 +22,7 @@ git pull
 #git pull
 
 # prep are vars
+export WIPE_DB=${WIPE_DB:-false}
 export POSTGRES_PORT="5433"
 #export DATABASE_URL="postgres://postgres:password@0.0.0.0:$POSTGRES_PORT/namada-indexer"
 export DATABASE_URL="postgres://postgres:password@postgres:$POSTGRES_PORT/namada-indexer"
@@ -84,7 +85,12 @@ cd $HOME/namada-indexer
 docker stop $(docker container ls --all | grep 'namada-indexer' | awk '{print $1}')
 docker container rm --force $(docker container ls --all | grep 'namada-indexer' | awk '{print $1}')
 
-docker compose -f $HOME/namada-indexer/docker-compose.yml down --volumes
+if [ "$WIPE_DB" = true ]; then
+    docker compose -f $HOME/namada-indexer/docker-compose.yml down --volumes    
+else
+    docker compose -f $HOME/namada-indexer/docker-compose.yml down
+fi
+
 POSTGRES_CONTAINER_ID=$(docker ps --filter "name=postgres" --filter "publish=${POSTGRES_PORT}" --format "{{.ID}}")
 if [ -n "$POSTGRES_CONTAINER_ID" ]; then
     echo "Stopping and removing 'postgres' container running on port ${POSTGRES_PORT}..."
@@ -100,9 +106,11 @@ docker image rm --force $(docker image ls --all | grep -E '^namada/.*-indexer.*$
 docker image rm --force $(docker image ls --all | grep '<none>' | awk '{print $3}')
 
 # prune all volumes (db data)
-docker volume prune -fa
+if [ "$WIPE_DB" = true ]; then
+    docker volume prune -fa
+fi
 
-# Copy the docker compose file for the db
+# Copy the docker compose file for the persistent db
 cp -f $HOME/namada-campfire/docker/compose/docker-compose-db.yml $HOME/namada-indexer/docker-compose-db.yml
 
 # build and start the containers
